Renamed restore_filename to filename
I renamed the restore_filename file within the sysfs to filename. The
reason for this rename was to be more general in the name. The filename
could be utilized to allow userspace to specify the name of the file to
create on a suspend action. It can also be used to direct the module
at a process during the restore action as well.