Added ability to write to file within kernel
Added an api to the lkm_utils.c file to support reading and writing to a file from
within the linux kernel. The control.c was updated to fix the bug that occurred when
a pid with greater than 4 digits was presented. The lkm_inspect.c file was update
to export the contents of the task_struct when the control application provides a
command of 1.