Began prototyping routine for deactivating an active process in the system
I renamed the LKM_TEST to LKM_INFO in the types.h file. Updated the lkm_inspect.c so that it
makes a call to the deactivate function that was added to the lkm_utils.c file. Began
prototyping the lkm_deactivate_pid routine to support the suspension of the process.