Rebuilt the inspection module to support a command structure
The inspection module used to make use of multiple file within the sysfs system.
I created a operation file descriptor and removed all the others. This operation
will now accept a structure that represents the operation that is being requested.
The operation allows for specifying the directory name for the files to be dumped.