Adding the kernel_config.h and the _generateHeader.py files
The kernel_config.h file contains the kernel configuration that the inspection module is
being loaded into. The inclusion of this file allows for the customization of the module
based on what modules are available. The _generateHeader.py file is a python script that
generates the kernel_config.h based on the .config file from the kernels configuration.