Removed the include for the kernel_config.h file from the export files.
The build process for a module already provides all the kernel defines. It
is not necessary to include any files in order to detect kernel configurations.
The _generateHeader.py file was left as it correctly converted the .config file
in the event that we still need the functionality going forward.